Removed obsolete code from projects/jobs.py . Reverted changes in chart-controller . Refactored code in deployments to prevent crashes from the changes .
