+++ components/chart-controller/controller/controller.py
@@ -105,11 +105,6 @@ class Controller:
         args = ['helm', '--kubeconfig', str(kubeconfig), 'delete', options['release']]
         status = subprocess.run(args, cwd=self.cwd)
 
-        # Delete k8s job used to allocate project resources
-        k8s_job_cmd = "kubectl get jobs | awk '/" + options['release'] +"-/{print $1}' | xargs kubectl delete job"
-        k8s_job_status = subprocess.run(k8s_job_cmd, shell=True, cwd=self.cwd)
-        print(k8s_job_status)
-
         return json.dumps({'helm': {'command': args, 'cwd': str(self.cwd), 'status': str(status)}})
 
     def update(self, options, chart):
